initial split of server into multiple files
